requirements.txt
fastapi
uvicorn
requests
google-generativeai
opencv-python-headless
numpy
pydantic
nest_asyncio